Moved solver functions to seperate file linearized_Usadel_solver_2D and updated both files
Created a separate file linearized_Usadel_solver_2D and move the solver functions to that file and removed them from 2D Usadel eq ghost. Added the boundary conditions as a function given to the solver so it handles different systems. Also fixed an error when returning the gradients from the Usadel solver. Added documentation to the functions in linearized_Usadel_solver_2D.