Added a file with magnetic field and fixed BC in matrix
Added a file where the magnetic field is included when we have a normal metal with a superconductor on all 4 sides and a magnetic field pointing through the normal metal. The BC when creating the matrix were fixed to have the correct sign and be Neuman type in all 4 directions.