Made a separate file for boundary conditions, added calculations for self consistent A-field and properly reproduced results
The boundary conditions were placed into a separate file to separate them from the calculations. Self consistency calculations for the A-field can now be done with a function calculating the contributions to the A-field from the currents in the system. The results from Amundsen et al. are now properly reproduced with the result respecting the discreet symmetries of the system.