Added a function for calculating the divergence free vector potential for a given B field
Added a function for calculating the divergence free vector potential for a given value of the B fieald and a given vector potential in the surrounding superconductor. Added the possibility to choose between Kuprianov-Lukichev and transparent boundary conditions as an input to the Usadel solver. Added a streamplot for the current. Added a lot of documentation to several of the functions, especially the boundary conditions.