Added functionality for thin films and new way of calcluating the A field
Added a solver for a thin film with the superconductor below where the boundary conditions at the bottom are integrated out to obtain an effective extra term. In addition added functions that allows for the calculation of the magnetic vector potential, A, by using Poincare's lemma to integrate the magnetic field.