Fixed sign error and calculation of B-field
Fixed two sign errors in the linearized Usadel equation. Also fixed the calculation of the B-field for the superconductor surrounding the normal metal when N is odd, though if N=4n+1 this still does not work. Finally fixed an error where division by the inverse distances was done instead of multiplication when calculating the induced A-field.